loan_management/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
future==0.18.3
mecab==0.996.3
pytz==2023.3
serializers==0.2.4
sqlparse==0.4.4
tzdata==2023.3
